# app/docs/openapi.yaml
openapi: 3.0.3
info:
  title: n8n Agent Dashboard API
  description: |
    API specification for the n8n Agent Dashboard - a self-feeding workflow automation platform.
    
    This API provides endpoints for:
    - Workflow generation using AI (Kimi 2.5, Gemini 2.5)
    - Workflow deployment via MCP Bridge or Browser Automation
    - Execution monitoring and management
    - Knowledge base operations
    
    ## Authentication
    
    All API endpoints require authentication via Bearer token:
    ```
    Authorization: Bearer your-api-token
    ```
    
    ## Deployment Modes
    
    - **MCP Bridge**: Direct API integration with n8n (recommended)
    - **Browser Automation**: Playwright-based browser automation (fallback)
  version: 1.0.0
  contact:
    name: n8n Agent Dashboard Team
    email: support@yourdomain.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://your-app.vercel.app/api
    description: Production server
  - url: http://localhost:5173/api
    description: Development server
security:
  - bearerAuth: []
...
paths:
...
  /workflows/{id}:
    get:
      summary: Get workflow by ID
      description: Retrieve a workflow from the database
      tags:
        - Workflows
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Workflow found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowRecord'
        '404':
          description: Workflow not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete workflow
      description: Delete a workflow from the database and n8n
      tags:
        - Workflows
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Workflow deleted
        '404':
          description: Workflow not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    WorkflowRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [draft, testing, deployed, failed, archived]
        version:
          type: integer
        n8n_workflow_id:
          type: string
        deployment_method:
          type: string
          enum: [mcp, browser-automation]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: object